refactor(frontend): implement MudPromptManager for managing prompt state and server output
