+++ frontend/src/app/features/terminal/index.ts
@@ -1,2 +1,3 @@
 export * from './models/escapes';
 export * from './mud-input.controller';
+export * from './mud-prompt.manager';
